use generic config observer
