+++ pkg/operator/staticpod_interface.go
@@ -52,3 +52,45 @@ func (c *staticPodOperatorClient) CurrentStatus() (operatorv1.OperatorStatus, er
 
 	return instance.Status.OperatorStatus, nil
 }
+
+func (c *staticPodOperatorClient) GetOperatorState() (*operatorv1.OperatorSpec, *operatorv1.OperatorStatus, string, error) {
+	instance, err := c.informers.Kubecontrollermanager().V1alpha1().KubeControllerManagerOperatorConfigs().Lister().Get("instance")
+	if err != nil {
+		return nil, nil, "", err
+	}
+
+	return &instance.Spec.OperatorSpec, &instance.Status.StaticPodOperatorStatus.OperatorStatus, instance.ResourceVersion, nil
+}
+
+func (c *staticPodOperatorClient) UpdateOperatorSpec(resourceVersion string, spec *operatorv1.OperatorSpec) (*operatorv1.OperatorSpec, string, error) {
+	original, err := c.informers.Kubecontrollermanager().V1alpha1().KubeControllerManagerOperatorConfigs().Lister().Get("instance")
+	if err != nil {
+		return nil, "", err
+	}
+	copy := original.DeepCopy()
+	copy.ResourceVersion = resourceVersion
+	copy.Spec.OperatorSpec = *spec
+
+	ret, err := c.client.KubeControllerManagerOperatorConfigs().Update(copy)
+	if err != nil {
+		return nil, "", err
+	}
+
+	return &ret.Spec.OperatorSpec, ret.ResourceVersion, nil
+}
+func (c *staticPodOperatorClient) UpdateOperatorStatus(resourceVersion string, status *operatorv1.OperatorStatus) (*operatorv1.OperatorStatus, string, error) {
+	original, err := c.informers.Kubecontrollermanager().V1alpha1().KubeControllerManagerOperatorConfigs().Lister().Get("instance")
+	if err != nil {
+		return nil, "", err
+	}
+	copy := original.DeepCopy()
+	copy.ResourceVersion = resourceVersion
+	copy.Status.StaticPodOperatorStatus.OperatorStatus = *status
+
+	ret, err := c.client.KubeControllerManagerOperatorConfigs().UpdateStatus(copy)
+	if err != nil {
+		return nil, "", err
+	}
+
+	return &ret.Status.StaticPodOperatorStatus.OperatorStatus, ret.ResourceVersion, nil
+}
